+++ packages/terra-docs/tests/wdio/placeholder-spec.js
@@ -1,7 +1,7 @@
 Terra.describeViewports('Placeholder', ['medium'], () => {
   describe('Default Placeholder', () => {
     it('checks accessibility', () => {
-      browser.url('/raw/tests/terra-docs/docs/placeholder');
+      browser.url('/raw/tests/cerner-terra-docs/docs/placeholder');
       Terra.validates.element('default');
     });
   });