+++ packages/terra-overlay/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 3.66.0 - (June 22, 2021)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.65.0 - (June 8, 2021)
+
+* Changed
+  * Updated `keycode-js` from `v2.0.1` to `v3.1.0`.
+
 ## 3.64.0 - (April 13, 2021)
 
 * Changed