+++ packages/terra-profile-image/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 3.40.0 - (July 6, 2021)
+
+* Changed
+  * Updated fallback image styles.
+
 ## 3.39.0 - (May 4, 2021)
 
 * Changed