+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 4.27.0 - (July 6, 2021)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.26.0 - (June 22, 2021)
+
+* Changed
+  * Minor dependency version bump
+
+## 4.25.0 - (June 8, 2021)
+
+* Changed
+  * Updated `keycode-js` from `v2.0.1` to `v3.1.0`.
+
 ## 4.24.0 - (April 6, 2021)
 
 * Changed